Refactor architecture, performance, and safety; fix CI failures
- Implement Builder pattern CRTP for fluent API.
- Adopt std::string_view for performance.
- Enforce SafeSpan usage in transport layer.
- Improve TcpServer shutdown stability.
- Fix CI compilation errors (python bindings, unit tests, performance tests).
- Fix warnings treated as errors (-Wsign-conversion, -Wunused-parameter, etc).
- Fix duplicate initialization in TcpServer constructor.
- Fix Python bindings on_bytes signature.

+++ test/performance/benchmark/test_memory_pool_growth.cc
@@ -80,7 +80,7 @@ TEST_F(MemoryPoolGrowthBenchmark, GrowthPerformance) {
     buffers.clear();
   }
 
-  double ms_total = total_duration.count() / 1000000.0;
+  double ms_total = static_cast<double>(total_duration.count()) / 1000000.0;
   double us_per_iter = (double)total_duration.count() / 1000.0 / iterations;
 
   std::cout << "Total Time: " << std::fixed << std::setprecision(3) << ms_total << " ms" << std::endl;
