+++ unilink/transport/tcp_server/tcp_server.cc
@@ -45,7 +45,7 @@ TcpServer::TcpServer(const config::TcpServerConfig& cfg)
       ioc_(concurrency::IoContextManager::instance().get_context()),
       acceptor_(nullptr),
       cfg_(cfg),
-      max_clients_(cfg.max_connections > 0 ? cfg.max_connections : 0),
+      max_clients_(cfg.max_connections > 0 ? static_cast<size_t>(cfg.max_connections) : 0),
       client_limit_enabled_(cfg.max_connections > 0),
       current_session_(nullptr) {
   // Create acceptor after all members are initialized
@@ -64,7 +64,7 @@ TcpServer::TcpServer(const config::TcpServerConfig& cfg, std::unique_ptr<interfa
       ioc_(ioc),
       acceptor_(std::move(acceptor)),
       cfg_(cfg),
-      max_clients_(cfg.max_connections > 0 ? cfg.max_connections : 0),
+      max_clients_(cfg.max_connections > 0 ? static_cast<size_t>(cfg.max_connections) : 0),
       client_limit_enabled_(cfg.max_connections > 0),
       current_session_(nullptr) {
   // Ensure acceptor is properly initialized
